# add benchmark for readLastLine method