This PR doesn't need to merge. It is just for comparing the benchmark result with the new implementation for readLastLine in PR #19 

Comparison:

```
Old way:
result for the last line of 76:
searchLogSuite.BenchmarkReadLastLine        10000            124423 ns/op
searchLogSuite.BenchmarkReadLastLine        10000            126135 ns/op

result for the last line of 76*2=152:
searchLogSuite.BenchmarkReadLastLine        10000            247836 ns/op
searchLogSuite.BenchmarkReadLastLine        10000            251958 ns/op

New way:
searchLogSuite.BenchmarkReadLastLines      1000000              1892 ns/op
searchLogSuite.BenchmarkReadLastLines      1000000              1920 ns/op
```

Conclusion: the new implementation is 65x (126135/1920) as fast as the old way when the last line is 76 long.