@baurine 2020-07-10
Close it after PR #19 is merged.